% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{nn_effect}
\alias{nn_effect}
\title{Effect of input}
\usage{
nn_effect(X, ind, W, q, val = rep(0, ncol(X)), length = 100, range = c(-3, 3))
}
\arguments{
\item{X}{input matrix}

\item{W}{weight vector}

\item{val}{value of input units}

\item{length}{length of vector to be calculated}

\item{range}{range of values to be computed}

\item{i}{Variable index}

\item{h}{number of hidden units}
}
\value{
Effect of variable
}
\description{
Effect of input
}
