% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_functions.R
\name{nn_fit}
\alias{nn_fit}
\title{Neural network fitting function using optimx}
\usage{
nn_fit(X, Y, W, q)
}
\arguments{
\item{X}{Input data}

\item{Y}{Output data}

\item{W}{Starting weight vector}

\item{q}{Number of hidden units}
}
\value{
Optimal weight vector
}
\description{
Neural network fitting function using optimx
}
