% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_model_sel_local.R
\name{nn_fit_tracks}
\alias{nn_fit_tracks}
\title{nn_fit_tracks}
\usage{
nn_fit_tracks(X, Y, q, W_mat, inf_crit = "BIC", unif = 3, ...)
}
\arguments{
\item{X}{input matrix}

\item{Y}{output vector}

\item{q}{number of hidden units}

\item{W_mat}{inital weights matrix}

\item{inf_crit}{information criterion (BIC only)}

\item{unif}{uniform distribution}
}
\value{
nn_fit_tracks
}
\description{
nn_fit_tracks
}
