% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{nn_variable_sel}
\alias{nn_variable_sel}
\title{Variable Selection}
\usage{
nn_variable_sel(X, Y, nn, n_iter, ...)
}
\arguments{
\item{X}{input matrix}

\item{Y}{output vector}

\item{nn}{output from nn_model_sel function}

\item{n_iter}{number of iteration}
}
\value{
Variable Selection
}
\description{
Variable Selection
}
