% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prof_likelihood.R
\name{prof_likelihood}
\alias{prof_likelihood}
\title{Profile likelihood: function to find optimal weights when fixing a parameter}
\usage{
prof_likelihood(X, Y, W, q, ind, val)
}
\arguments{
\item{X}{Data}

\item{Y}{Response}

\item{W}{Weight vector}

\item{q}{number of hidden units}

\item{ind}{index of weight for profiling}

\item{val}{value for profiling}
}
\value{
Profile likelihood weights
}
\description{
Profile likelihood: function to find optimal weights when fixing a parameter
}
