% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{var_imp}
\alias{var_imp}
\title{Variable Importance}
\usage{
var_imp(
  X,
  Y,
  ind,
  n_iter,
  W = NULL,
  W_mat = NULL,
  q,
  unif = 1,
  inf_crit = "BIC",
  ...
)
}
\arguments{
\item{X}{input matrix}

\item{Y}{output vector}

\item{ind}{Variable index}

\item{n_iter}{number of iteration}

\item{W}{weight vector}

\item{q}{number of hidden units}

\item{unif}{uniform distribution}
}
\value{
Variable importance
}
\description{
Variable Importance
}
