% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_importance.R
\name{nn_variable_sel}
\alias{nn_variable_sel}
\title{Variable Selection}
\usage{
nn_variable_sel(X, Y, n_iter, q = NULL, nn = NULL, unif = 3, ...)
}
\arguments{
\item{X}{input matrix}

\item{Y}{output vector}

\item{n_iter}{number of iteration}

\item{nn}{output from nn_model_sel function}
}
\value{
Variable Selection
}
\description{
Variable Selection
}
